+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - jrl_release: add package.xml
 - jrl_check_python_module: add version handling by @ahoarau
+- jrl_option: fix condition handling on reconfigure and support parent project normal variables (CMP0077)
 
 ## [2.2.4] - 2026-08-07